Move InputFingerprinter to its own package
